final project

+++ medRecords/src/main/java/medrecords/domain/impl/DoctorImpl.java
@@ -2,15 +2,21 @@ package medrecords.domain.impl;
 
 import medrecords.dao.DoctorRepository;
 import medrecords.domain.DoctorService;
+import medrecords.exception.InvalidEntityDataException;
 import medrecords.exception.NonexisitingEntityException;
 import medrecords.model.Doctor;
 import medrecords.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
+
+
 public class DoctorImpl implements DoctorService {
 private DoctorRepository doctorRepository;
 
@@ -46,9 +52,21 @@ private DoctorRepository doctorRepository;
     @Override
     public Doctor deleteDoctorById(Long id) {
         var old = findDoctorById(id);
-        doctorRepository.deleteById(id);
-        return old;
+        if(old.getAppointments().isEmpty() && old.getMedicalRecords().isEmpty() && old.getPrescriptions().isEmpty()){
+            doctorRepository.deleteById(id);
+            return old;
+        }
+        else {
+            throw new InvalidEntityDataException(String.format("There are documents associated with this doctor."));
+        }
+
     }
+
+    @Override
+    public List<Doctor> findDoctorBySpeciality(String speciality) {
+        return doctorRepository.findBySpecialityContainingIgnoreCase(speciality);
+    }
+
 
     @Override
     public long countDoctors() {
